requirements.txt
pillow==10.3.0
pix2text==1.1.0.1
Flask==3.0.3
openai==1.25.1
python-dotenv==1.0.1